requirements.txt
numpy==2.4.2
pandas==3.0.1
pyarrow==23.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
